# internal/app/api/core/assets/doc/v0_swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "WireGuard Portal API - UI Endpoints",
        "title": "WireGuard Portal SPA-UI API",
        "contact": {
            "name": "WireGuard Portal Developers",
            "url": "https://github.com/h44z/wg-portal"
        },
        "version": "0.0"
    },
    "basePath": "/api/v0",
    "paths": {
...
        "/config/frontend.js": {
            "get": {
                "produces": [
                    "text/javascript"
                ],
                "tags": [
                    "Configuration"
                ],
                "summary": "Get the dynamic frontend configuration javascript.",
                "operationId": "config_handleConfigJsGet",
                "responses": {
                    "200": {
                        "description": "The JavaScript contents",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
...
    },
...
}